lsfd: use NAME column to show cooked file names
This change splits the NAME column
into cooked NAME column and raw KNAME column.

KNAME represents the type information extracted from /proc/$pid/fd or
/proc/$pid/map_files. NAME is the same as KNAME but humans friendly
and talkative if a special handler is implemented for the type of the
file.

Currently only pidfd has such a handler.

Below an example output:

    $ ./lsfd -o TYPE,STTYPE,KNAME,NAME -Q '(TYPE == "pidfd")'
     TYPE STTYPE KNAME              NAME
    pidfd   UNKN anon_inode:[pidfd] pid=4661 comm=dbus-broker nspid=4661
    pidfd   UNKN anon_inode:[pidfd] pid=4924 comm=dbus-broker nspid=4924

Signed-off-by: Masatake YAMATO <yamato@redhat.com>