lsfd: show pid, comm, and nspid of pidfd in PIDFD.{PID,COMM,NSPID} columns
The values shown in the new columns are already shown in NAME column.
Showing these values together in the NAME column helps users
understand the pidfds opened quickly. However, the showing them
together is not suitable for machine processing. The newly introduced
columns are for machine processing.

Below an example output:

    $ ./lsfd -o COMMAND,TYPE,PIDFD.COMM,PIDFD.PID -Q '(TYPE == "pidfd")'
    COMMAND          TYPE PIDFD.COMM  PIDFD.PID
    dbus-broker-lau pidfd dbus-broker      4661
    dbus-broker-lau pidfd dbus-broker      4924

Signed-off-by: Masatake YAMATO <yamato@redhat.com>