[su] stop parsing arguments after - and --
Currently su will try to parse all flags, even those after "-", which should be
passed on to the spawned program. However, getopt will try to interpret all
flags and mostly these do not belong to su and cause an error.
Thus we set argc to the position of "-" or "--" and ensure that the rest is
never interpreted.