fadvise: new command wrapping posix_fadvise
For measuring and monitoring the performance aspect of a command,
people may want to drop the page caches related to specified files.

In some cases, writing 1 to /proc/sys/vm/drop_caches is overkill.  The
fadvise command can be used for dropping page caches related to
specified files; the impact on a system is much smaller than
/proc/sys/vm/drop_caches.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>