lsfd: introduce SOCK. column namespace
lsfd has many columns already. We will add more in the future.
e.g. --help output is too long.

Groping columns is one of the ways to improve the usability of lsfd.
As the first step, this change introduces the "SOCK." column namespace.

Conceptually SOCKNETNS, SOCKSTATE, and SOCKTYPE migrate to the
namespace.  At the implementation level, they are renamed to
SOCK.NETNS, SOCK.STATE, and SOCK.TYPE.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>