procfs: get_stat_nth: handle braces in process name
In procfs_process_get_stat_nth() we skip to the end of the process name
by looking for the ')' after it.
However if the process name itself contains a ')' then find that instead
of the correct one.

By searching the ')' from the end of the file we can make sure to always
find the correct one.

Link: https://www.openwall.com/lists/oss-security/2022/12/21/6