tests: (lsfd) adjust the output for unix stream sockets
lsfd changes the output of NAME column for a unix stream socket
whether the kernel reports it is a "UNIX-STREAM" socket or a "UNIX"
socket. Rather older version of kernels report the socket "UNIX"

For "UNIX", lsfd appends "type=stream" to the NAME column.
This caused a failure in test cases. To avoid the failure,
this change strips "type=stream" from the output before
comparing the output strings and expected strings.

See also the commit log of a60ac11fe0087584f37329b4733edfade3452c64.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>