test: (lsfd) verify the current groups before making a PING socket
For making a PING socket, the groups of the process creating
the socket must be in the range specified in
/proc/sys/net/ipv4/ping_group_range.

With this change, mkfds-ping test case verifies the groups before
making a PING socket with test_mkfds helper. If the current groups are
out of the range, skip the testing.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>